Add explicit casts to int to enforce contract between method signature and data types. Helps avoid deprecation warnings in Numpy random chooser methods.

+++ urbansim/models/dcm.py
@@ -1822,7 +1822,7 @@ class SegmentedMNLDiscreteChoiceModel(DiscreteChoiceModel):
                 (len(alternatives), len(choosers)))
             idxes = np.random.choice(
                 alternatives.index,
-                size=np.floor(len(choosers) * alternative_ratio),
+                size=int(np.floor(len(choosers) * alternative_ratio)),
                 replace=False)
             alternatives = alternatives.loc[idxes]
             logger.info(
